requirements.txt
streamlit
PyMuPDF
Pillow
pytesseract
pandas
openpyxl
